x/bridge/keeper/query_last_injected_txs_nonce.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/fuel-infrastructure/fuel-sequencer/x/bridge/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) LastInjectedTxsNonce(goCtx context.Context, req *types.QueryGetLastInjectedTxsNonceRequest) (*types.QueryGetLastInjectedTxsNonceResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(goCtx)

	val, found := k.GetLastInjectedTxsNonce(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetLastInjectedTxsNonceResponse{Nonce: val.String()}, nil
}
